+++ app/redux/services/base/base-api.service.ts
@@ -9,23 +9,28 @@ const REFRESH_TOKEN: string = 'refresh_token';
 const SEARCH_PARAMS: any = 'search_params';
 
 const getSearchParams = () => {
-    return localStorage.getItem(SEARCH_PARAMS);
+    if (typeof window !== 'undefined')
+        return localStorage.getItem(SEARCH_PARAMS);
 }
 
 const setSearchParams = (value: any) => {
-    localStorage.setItem(SEARCH_PARAMS, value);
+    if (typeof window !== 'undefined')
+        localStorage.setItem(SEARCH_PARAMS, value);
 }
 
 const removeSearchParams = () => {
-    localStorage.removeItem(SEARCH_PARAMS);
+    if (typeof window !== 'undefined')
+        localStorage.removeItem(SEARCH_PARAMS);
 }
 
 const getEmpId = () => {
-    return localStorage.getItem(EMPID);
+    if (typeof window !== 'undefined')
+        return localStorage.getItem(EMPID);
 }
 
 const setEmpId = (empid: string): void => {
-    localStorage.setItem(EMPID, empid);
+    if (typeof window !== 'undefined')
+        localStorage.setItem(EMPID, empid);
 }
 
 const getUserInfo = () => {
@@ -43,28 +48,36 @@ const getUserInfo = () => {
 }
 
 const setUserInfo = (info: string): void => {
-    localStorage.setItem(USER, info);
+    if (typeof window !== 'undefined')
+        localStorage.setItem(USER, info);
 }
 
 const getToken = () => {
-    return localStorage.getItem(TOKEN_NAME);
+    if (typeof window !== 'undefined')
+        return localStorage.getItem(TOKEN_NAME);
 }
 
 const setToken = (token: string) => {
-    localStorage.setItem(TOKEN_NAME, token);
+    if (typeof window !== 'undefined')
+        localStorage.setItem(TOKEN_NAME, token);
 }
 
 const getRefreshToken = () => {
-    return localStorage.getItem(REFRESH_TOKEN);
+    if (typeof window !== 'undefined')
+        return localStorage.getItem(REFRESH_TOKEN);
 }
 
 const setRefreshToken = (token: string) => {
-    localStorage.setItem(REFRESH_TOKEN, token);
+    if (typeof window !== 'undefined')
+        localStorage.setItem(REFRESH_TOKEN, token);
 }
 
 const removeToken = () => {
-    localStorage.removeItem(TOKEN_NAME);
-    localStorage.removeItem(REFRESH_TOKEN);
+    if (typeof window !== 'undefined') {
+        localStorage.removeItem(TOKEN_NAME);
+        localStorage.removeItem(REFRESH_TOKEN);
+    }
+        
 }
 
 const ssoLogin = () => {
@@ -72,7 +85,7 @@ const ssoLogin = () => {
     // admin 에서는 직접 로그인 막음
     // window.location.href = `${process.env.NEXT_PUBLIC_SSO_AUTH}`;
 
-    localStorage.clear();
+    if (typeof window !== 'undefined')  localStorage.clear();
 
     const LOGOUT_URL = window.location.protocol + '//' + window.location.host + '/auth/logout'
     window.location.href = LOGOUT_URL;
@@ -84,7 +97,7 @@ const ssoLogout = () => {
     // localStorage.clear();
     // window.location.href = `${process.env.NEXT_PUBLIC_SSO_LOGOUT}`;
 
-    localStorage.clear();
+    if (typeof window !== 'undefined')  localStorage.clear();
 
     const LOGOUT_URL = window.location.protocol + '//' + window.location.host + '/auth/logout'
     window.location.href = LOGOUT_URL;
